sansbox/tracer/lib.rs: just fixed a typo.
rust still don't allow default argument and i used it to improve user experience.
when reading string, we want to constrain to length of string. But we also want user to have the ability to change that constrain.
the limitation is now 256 bytes, and since rust don't allow default argument, it 256 forever :)

Signed-off-by: peter <PeterRong96@gmail.com>