sandbox/safe-box/src/*: permission.rs changed.
We need to re-consider out open/ execve policy again and carefully.
Now that open a file outside /home can be considered to be allowed. Afterall, if the programme attempts something bad, os will stop it.
However, we may not simply reject all files that insde /home but outside our safe zone. We need user to tell what's safe and what's not.

Signed-off-by: peter <PeterRong96@gmail.com>